Disable LLVM ABI breaking checks by default when building swift
The LLVM ABI breaking checks require that all uses of llvm's ADT library must also link libSupport. Parts of the Swift standard library use ADT without Support intentionally. To make this work we must disable the ABI breaking checks when building swift.

This patch changes the default value of the option that controls the ABI breaking checks to disable the checks by default if the swift source directory is present in the LLVM build tree.