X86: remove R12 from CSR on Windows x64 SwiftCC
R12 is used for the SwiftError parameter.  It is no longer a CSR as it
is used for transfer the SwiftError, and the caller must preserve it if
they need to.

Cherry picked by aschwaighofer for swifterror support on non darwin platforms.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@314165 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit e509babaee14a6e05086404b405acc6cfe5c6349)