master-next: Temporary workaround for llvm r315852.
The change in llvm r315852 is causing an assertion failure for Swift.
I've notified the author of that change and am hoping for a fix. In the
meantime, here is a workaround. I'm doing this instead of reverting the
offending commit from swift-llvm's upstream-with-swift branch in anticipation
of a merge conflict from the hoped-for fix. I'll remove this workaround when
there is a real fix.

rdar://problem/35017353