[dsymutil] Implement the --threads option
This patch adds the --threads option to dsymutil to process
architectures in parallel. The feature is already present in the version
distributed with Xcode, but was not yet upstreamed.

This is NFC as far as the linking behavior is concerned. As threads are
used automatically, the current tests cover the change in
implementation.

Differential revision: https://reviews.llvm.org/D39355

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@316999 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit ac330efcdf3a70ca18b65d3dbfa398b9fff68149)

Conflicts:
	tools/dsymutil/DwarfLinker.cpp
(cherry picked from commit 200e73656a9ccead3efb19fc2b356ca47d492b57)